Task 3

+++ 0x05-python-exceptions/3-safe_print_division.py
@@ -1,7 +1,7 @@
 #!/usr/bin/python3
 def safe_print_division(a, b):
     try:
-        return a / b
+        return float(a / b)
     except ZeroDivisionError:
         return None
     finally:
